Mobilender/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.19
django-filter==2.4.0
djangorestframework==3.12.2
django-rest-swagger==2.2.0
idna==2.10
importlib-metadata==3.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
pip==20.1.1
pluggy==0.13.1
psycopg2==2.8.6 
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-django==4.1.0
pytz==2021.1
requests==2.25.1
setuptools==47.1.0
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
zipp==3.4.0
django-environ==0.4.5
